refactor: Split import sdk in import specific and not import specific parts
To prepare for the possibility of future sdks similar to this import sdk, we apply some changes on the namespace structure:
- Namespaces now reflect that import is just a part of the piweb sdk by moving from Zeiss.PiWeb.Import.Sdk to Zeiss.PiWeb.Sdk.Import. This leaves room for additional sdk types.
- There is now a Common namespace below Zeiss.PiWeb.Sdk for anything that may be shared between different sdk types. In the future, when necessary, we can move this namespace to its own assembly referenced by all sdks.
- Several components that are not import specific are moved to the common namespace.

For now this means that the root namespace of the import sdk assembly is Zeiss.PiWeb.Sdk and not Zeiss.PiWeb.Sdk.Import. However, when we extract the common assembly in the future, we can use Zeiss.PiWeb.Sdk.Import as root namespace again.